move pipoinputs
